Move to component, fix checking not permitted things on check all
